ml-models/bacilli_detr/requirements.txt
torch
pillow
numpy
torchvision
fastapi 
typer
uvicorn
packaging
scipy